Move Login contract to the shared module
