+++ platform-shared/src/commonMain/kotlin/io/vitalir/kotlinhub/shared/feature/user/LoginResponse.kt
@@ -1,0 +1,10 @@
+package io.vitalir.kotlinhub.shared.feature.user
+
+import kotlinx.serialization.Serializable
+import kotlin.js.JsExport
+
+@Serializable
+@JsExport
+class LoginResponse(
+    val token: String,
+)
