+++ server/application/src/main/kotlin/io/vitalir/kotlinhub/server/app/feature/user/routes/get/GetUserByLoginResponse.kt
@@ -1,6 +1,6 @@
 package io.vitalir.kotlinhub.server.app.feature.user.routes.get
 
-import io.vitalir.kotlinhub.server.app.feature.user.routes.common.ApiUser
+import io.vitalir.kotlinhub.shared.feature.user.ApiUser
 import kotlinx.serialization.Serializable
 
 @Serializable
