--- server/application/src/main/kotlin/io/vitalir/kotlinhub/server/app/feature/user/routes/login/LoginResponse.kt
@@ -1,8 +1,0 @@
-package io.vitalir.kotlinhub.server.app.feature.user.routes.login
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-internal data class LoginResponse(
-    val token: String,
-)
